+++ NEWS.md
@@ -2,6 +2,7 @@
 * Export from replicated Campsis model too slow, revise implementation #99
 * OMEGA/SIGMA matrix export too slow, revise implementation #100
 * Sampling from inverse Wishart/Chi-squared: degrees of freedom should be a vector #101
+* Implement show method on replicated Campsis model #103
 
 # campsismod 1.2.0